Tests for BasicBoard, reset Model from board
Allow a BoardModel to be reset from a BasicBoard, to copy the state.
This will be useful for initializing the model from a simple board,
created by some generator.  The other reset is useful for serializing
an in-flight game from a save state.

Updated the ascii-text test LevelCreator to initialize a BoardModel
*or* a BasicBoard, and take the object to initialize in the finished()
call instead of storing it in the constructor.  It also uses functions
in CellState rather than duplicating the switch logic.

New test file for BasicBoard.