Sketching out solver ideas.  Add Illum cellstate.
New solver directory, still just a scratchpad.

Made BoardModel default constructible, and BasicBoard moveable.

Added an Illum state to cellstate, to indicate a cell is illuminated.
While this is just a side effect of a bulb and not a playable move
state, the solver will benefit from knowing the square is no longer
playable.  Possibly could have put a "Mark" in the cell instead, but
this seemed easier to think about.  Maybe I'll revert it to a mark
later.