Change all row/col api to a Coord object
In addition to a Coord class, also changed StateChangeHandler to take
2 states instead of one: the previous and the next state.  It was kind
of messy before, since it only took the new state, but wasn't clear
what it was before.  Removes changed into the empty state, but
published the state of what it used to be, but adds published what
they were changing into.  This inconsistency caused issues.  Also, and
related, apply_move was a little wonky, in that it received one state
and passed it out, and set the tile.  Now it takes both and what is
sent is less ambiguous (or requiring more special casing.)

StartGame and ResetGame actions still are given states but they are
passed as Empty,Empty (from/to) since state doesn't really apply to
these.  Still hesitating on making more than one callback interface
function for state changes, but we'll see :)