New AnalysisBoard, plus visitor logic changes
Analysis board record is used for solving the game from any starting
position.  It doesn't work entirely yet, but it makes a copy of the
board for each move, plus some stats about the board.  This makes it
easy to search for moves to play.  (Well, it should.)  The boards are
so small that I'm not especially worried about efficiency.

visiting a row or column did not make sense after hitting a wall to
continue, so after any those row/col visitors visit a wall, the loop
breaks.

For the visit_rows_cols_outwards(), it will visit _each_ direction
until hitting a wall or until visitor returns false.  This is the only
visit* function where you can continue to visit cells after returning
false, but I think it makes more sense this way.  Before it was
indeterminate what was visited since direction is not specified, and
most reasons to stop involved hitting a wall anyway.

Also, due to changes above and failing tests, started to add some
DebugLog logic based on "fmt" library, so may be adding some
specializations to core classes soon for that--and because it's
fun!