speculation tree beginnings
This is building off an idea to find the simplest possible move,
rather than just any old move.  The approach is to speculate on every
possible move, first with "depth" of 1.  If we find a move, we are
done.  Play it.  If not, start over, searching for moves with depth of
2, etc.  This way we won't take a more complex solution at point A
when there's a simpler move that starts at point B.

First pass will be pure brute force.

Once that works, I'll keep state along the way to avoid re-visiting
known dead ends, and only explore the paths that still have potential.