Big checkin, compiles, not all tests pass
Added new idea of AnnotatedMove, which holds a SingleMove, but also
includes information on: what context decided this move, what the
motive was (under speculation, a forced move, or a followup to a
previous move (like after playing the final bulb around a wall,
filling remaining faces with marks.)

Also made utils directory and moved debuglog and new EnumUtils to it.
EnumUtils will declare a fmt-style formatter for enums, and provides
operator+ for the prefix version of to_underlying--which also isn't
yet available until c++23.

Solution gains interface to be more "board-like", instead of having to
access the underlying board to do so many things.

New solver solves several tests, but the significant onces still elude
it.  TBD.