Reduced trivial-move search code, fix bugs
Got rid of over half the trivial search code. Previously I had two
versions of each algorithm, one to play a single move and stop
searching, and the other to add _all_ moves into a vector.  Common
code was factored out.

However, this was still overly repetitive and complex, and so I
decided to gut the "single move" option and only insert moves into a
vector.  Now considering making a callback with each move so caller
can decide what to do with it.  Anyway.  Much less code to read,
debug, manage, and it's more consistently used.

Updated tests and found I was adding a bulb instead of a mark to empty
faces on satisfied walls.  Tests are good.  Fixed.  Some other tests
still do not pass, however.  Steps.  This is a big refactor.