Improve the previous fix for erasing children from map
The bug was that erasing the current element invalidates the iterator,
and the iteration mechanism was a ranged for loop.  So incrementing
after erasing was invalid.  Initial fix was just to use erase_if after
the loop, but changing to a regular for loop I can increment the
iterator in the common case, but use the result of
unordered_map::erase() for that path.  Saves a pass over the map.