Better profiling, new algorithm, MUCH faster
Speculation is vastly simpler, logic is smaller, and runs
significantly faster too.  What took 1.8 billion cell visitations now
takes only about 2 million, and while it took 15 second to complete
before, now it takes about 14 milliseconds.

WAS:

    visit_rows_cols_outward_counter       145327019
    visit_perp_counter                    163757884
    visit_col_down_counter                405296683
    visit_col_up_counter                  217630743
    visit_row_right_counter               379401591
    visit_row_left_counter                236781179
    visit_empty_counter                          84
    visit_adjacent_counter                318489726
    visit_board_counter                     8042912
    visit_cell_counter                   1811981637

    (ran in 18048 ms)

NOW:

    visit_rows_cols_outward_counter           71667
    visit_perp_counter                        21912
    visit_col_down_counter                   104803
    visit_col_up_counter                      85903
    visit_row_right_counter                   97844
    visit_row_left_counter                    79343
    visit_empty_counter                           2
    visit_adjacent_counter                    90657
    visit_board_counter                        2984
    visit_cell_counter                      2041876
    (ran in 28 ms)

Difference: No longer creates a tree, and reuses vectors to reduce
memory allocation.

Algorithm:

First, play all forced moves until none remain.
These are:

* a wall that requires N bulbs next to it has exactly N open faces.
  So they MUST be bulbs.

* a wall that already has its deps satisfied cannot have more, so add
  marks to any open faces of that wall.

* an empty space that cannot see another empty space in any direction
  must have a bulb.

* a mark that can only see a single empty space in every direction
  must have a bulb in that empty space in order to be illuminated.

* multiple empty spaces in the same row or column that form ambiguous
  solutions cannot be part of the solution, so must be marks:
  That is, on a board like below (0 are walls with no deps, . are empty spaces)

  000000.
  .......
  000000.

  The first 6 empty spaces in the middle row must be marks, because
  playing a bulb in one is equally good as playing a bulb in another,
  and that makes multiple solutions possible, which is not allowed, so
  this forces these X's:

  000000.
  XXXXXX.
  000000.

A followup move would then be forced bulb at the remaining empty space
in the middle row, since it's the only way to illuminate the marks.
It also happens to solve the rest of the board too.

After ALL of the above forced moves, and subsequent new forced moves
and so on are played out, either the board is solved, or there are no
more forced moves and we must speculate... that is, play "What If".

When speculating, we "imagine" playing a bulb in an empty cell and see
what happens by playing out the forced moves (if any) Eventually we settle into one of 3 states:
1) the play lead to a contradiction.
2) the play lead to a solution.
3) the sequence of forced moves played out, but did not solve and did
not lead to a contradiction.  (A dead-end.)

To Speculate:

For each empty cell on the board, create two speculation contexts for
it, which copies the current board with either either a mark or a bulb
played in it.  Then we loop across all the contexts, doing a single
iteration of "find forced (trivial) moves" and applying _all of them_.

Once any context reaches a contradiction, we complete evaluating all
the contexts, and apply th context(s) that arrived at contradictions
that iteration.

If no contradictions are found, dead ends are removed and we do
another iteration of search, etc.

Once contradictions are found, we stop searching for more and have
learned what we can from this speculation.