More efficiently add_bulbs to the board
when playing a bulb to the board, it would emit rays in all 4
directions.  For each cell along the way, it would scan all adjacent
cells looking for walls.  However, half of those scans were along the
"ray" being cast, and was a waste.  I added a new visit_adj_flanks to
allow adjacent scanning of just the left and right, and reduce some
redundant scanning.

Also added game5 to speculation tests.