new optimized find_isolated_cells
While I've probably reached a point where I can quit worrying about
performance, it's fun to squeeze out as much as I can.  The previous
iteration was a good improvement but I think it made the code hard to
read, and I thought I could do better.

Then I realized a nice way to do both: improve performance more, and
also make code simpler.  The new "find isolated" algorithm is more
complex than the original version but less complex than the first
optimized redo of it.

Now, it scans the board and builds two caches, one counting empty
cells in contiguous spans of "translucent cells" in rows, and the
other in columns.  Translucent in this case is "light can pass through
it": empty, marks, and illuminated cells.

For each span, it stores the count of empties, and then walks the
board, looking up the corresponding coungs for the row-span and
col-span it's in, and that's all it needs to determine what to do.

* If the current cell is Empty, and the sum of row+col spans is 2,
  then it is the empty counted in each, and so it sees no visible
  neighbors and must be a bulb.

* If the current cell is a Mark, and the sum of row+col spans is 1,
  then that 1 empty cell MUST be a bulb because nothing else can
  illuminate the mark.

* If the current cell is a Mark, and the sum of row+col spans is 0,
  the board is in error, since no cell can illuminate the mark.