CellState bitwise groups, and Solver state
All of the predicate functions for categorizing CellStates now have
their bitsets moved out to named constants in a namespace, and the
functions use those constants.  This makes them available to other
code that, for some reason, may want the bitset rather than the
function for comparison.

Also, Solver didn't set solved status or error status if it was
invoked in that stat, and immediately returns if it's asked to solve a
solved board, rather than asserting.