Update position board interface
Main improvement: add set_cell().  Without this, you could only
add_bulb or add_mark.  That sufficed for solvers which only made
"user-plays" onto the board, but for level generation itself it was
not enough, as changing the walls and situations themselves seems
important.

set_cell() is complicated in comparison, because it could break
assumptions by the current state, modifying the underlying board
without updating the position rules.  So it was left out deliberately.
But now it is added, and first, checks for some simple cases (where
set_cell can be translated into a known add_bulb or add_mark call (and
forwards calls to those instead) or it noops, or does something
trivial, in which case it just returns "success".

But for other cases, it is pedantic and simply edits the cell, and
recomputes the state of the board _from the beginning_ by playing out
all the moves again in an off-to-the-side copy, to learn what the
state should be now.

And there are some "not simple and not unknown changes" now, because
this change adds "set_wall" which has interesting complexity due to
blocking some illumination requiring removing some
illumination--unless it's also illuminated from a cross-beam from
another bulb.

For situations even more complex, it devers to modifying the board and
then restarting... adding the walls first and then adding the bulbs
and marks and recomputing the error state, etc.

For efficiency and "atomic grouping of changes" so that we don't judge
intermediate states, we may want to do a batch job of changing several
cells WITHOUT recomputing the state until the batch changes are all
applied, and then do a single FINAL evaluation.  To control this there
is a SetCellPolicy, which by default will not reevaluate the trivial
changes, but will reevaluate the settings it does not know.  However,
through the policy, one can prevent reevaluation even though the board
doesn't necessarily have the proper game state anymore (a la the batch
described above), and can also force an reevaluation even for trivial
moves (so we can guarantee the "last" change in a batch causes
reevaluation.

Other changes:

Added a reset(height, width) function to position board.  Previously
you could not reset it to an empty or resized board.

Added some missing visit_* forwarding functions, for completeness.

Added some type aliases to reduce ns-qualifier noise inside class.

Added new test file, PositionBoardTests, which is mostly migrated out
of and renamed AnalysisBoardTest functions (which should have been
done when the position board was itself split out of the analysis
board, but... hobby project and one-man crew working on this and
all... :)  But it includes some new tests for add_wall and for set_cell.

Also added some more tests for the new position board functionality.

Added macro to define BIT FUNCTIONALITY for enums operator &, |, &=,
|=, and contains_all.

Added bitwise functionality to Direction, utilizing the enum macro.

BasicBoard visit_rows_cols_outwards now takes a Direction bitset
indicating which directions that the use is interested in visiting
(default is all 4)