BoardModel reset gets policy arg
reset used to always copy the whole board, including user-played
moves.  It did try to defer user moves but there was a bug where it
copied the whole board and then applied user moves on top of it.  Net
effect, we did not discover new errors because they were not seen from
a valid board one at a time.  Stop copying the board initially from the argument, but copy cell-by-cell according to the newly added policy: copy the entire board or just the walls, leaving it in an unplayed, pristine state.

Also, use is_dynamic_entity instad of an explicit bitwise test, for
readability.