Adding a (bad) level generator
It works, it's clunky, I'm not proud of its design, but I can generate
thousands of valid levels with it.  I'm not sure if they are
interesting or have any difficulty, but that's iterable in the future.
It is enough for now so I can work on the GUI side of things with some
playable levels.  Yeehaw!