Begin refactor of solver code
I'd like to break the solver speculation logic into more pieces.  For
level generation I think speculating for just a single desired move
could utilize the solver logic, but currently it builds all possible
moves of bulbs and marks and speculates on each.  That speculation
could be a separate function, so I could apply it to just one.