Refactor speculation further
This breaks the initialization and the
"add_speculation_context_for_move" to the speculation context out to
their own functions.  The main init function loops calling
add_speculation_context_for_move, and now we can invoke
add_speculation_context_for_move in isolation, which is expected to be
useful in level generation.